[NFC] Update AOCX tests to improve shared source usage
