+++ sycl/test/fpga_tests/fpga_aocx_win.cpp
@@ -11,50 +11,14 @@
 
 /// E2E test for AOCX creation/use/run for FPGA
 // Produce an archive with device (AOCX) image
-// RUN: %clang_cl -fsycl -fintelfpga -fsycl-link=image -DDEVICE_PART %s -o %t_image_cl.a
+// RUN: %clang_cl -fsycl -fintelfpga -fsycl-link=image %S/Inputs/fpga_device.cpp -o %t_image.lib
 // Produce a host object
-// RUN: %clang_cl -fsycl -fintelfpga -DHOST_PART %s -c -o %t.o
+// RUN: %clang_cl -fsycl -fintelfpga -DHOST_PART %S/Inputs/fpga_host.cpp -c -o %t.obj
 
 // AOCX with source
-// RUN: %clang_cl -fsycl -fintelfpga -DHOST_PART %s %t_image_cl.a -o %t_aocx_src.out
+// RUN: %clang_cl -fsycl -fintelfpga -DHOST_PART %S/Inputs/fpga_host.cpp %t_image.lib -o %t_aocx_src.out
 // AOCX with object
-// RUN: %clang_cl -fsycl -fintelfpga %t.o %t_image_cl.a -o %t_aocx_obj.out
+// RUN: %clang_cl -fsycl -fintelfpga %t.obj %t_image.lib -o %t_aocx_obj.out
 //
 // RUN: env SYCL_DEVICE_TYPE=ACC %t_aocx_src.out
 // RUN: env SYCL_DEVICE_TYPE=ACC %t_aocx_obj.out
-
-#include "CL/sycl.hpp"
-#include <iostream>
-
-using namespace cl::sycl;
-
-#ifdef DEVICE_PART
-
-const double big[] = {3, 2, 1, 5, 6, 7};
-void foo(double &result, queue q, int x) {
-  buffer<double> buf(&result, 1);
-  buffer<double, 1> big_buf(big, sizeof(big) / sizeof(double));
-  q.submit([&](handler &cgh) {
-    auto acc = buf.get_access<access::mode::discard_write>(cgh);
-    auto big_acc = big_buf.get_access<access::mode::read>(cgh);
-    cgh.single_task<class test>([=]() {
-      acc[0] = big_acc[x];
-    });
-  });
-}
-
-#endif // DEVICE_PART
-
-#ifdef HOST_PART
-
-void foo(double &, queue q, int x);
-
-int main(void) {
-  queue q(accelerator_selector{});
-
-  double result;
-  foo(result, q, 3);
-  std::cout << "Result: " << result << "\n";
-}
-
-#endif // HOST_PART
